+++ main.go
@@ -1,31 +1,12 @@
 package main
 
 import (
-	server "github.com/Likhil45/E-Commerce/Server"
+	controller "github.com/Likhil45/E-Commerce/Controller"
 	"github.com/Likhil45/E-Commerce/database"
 )
 
 func main() {
 	database.ConnectToMongoDB()
-	server.StartServer()
+	controller.StartServer()
 
-	// coll := client.Database("sample_mflix").Collection("movies")
-	// title := "Back to the Future"
-
-	// var result bson.M
-	// err = coll.FindOne(context.TODO(), bson.D{{"title", title}}).
-	// 	Decode(&result)
-	// if err == mongo.ErrNoDocuments {
-	// 	fmt.Printf("No document was found with the title %s\n", title)
-	// 	return
-	// }
-	// if err != nil {
-	// 	panic(err)
-	// }
-
-	// jsonData, err := json.MarshalIndent(result, "", "    ")
-	// if err != nil {
-	// 	panic(err)
-	// }
-	// fmt.Printf("%s\n", jsonData)
 }
